services/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "1.0.0"
description = "ProStream auth-service (FastAPI, SQLAlchemy 2.0, Pydantic 2)"
authors = ["Abdumalik-ProDev <abdumalikbakhtiyorovwork@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"

fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}

# DB / ORM
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.13.1"

# Validation & settings
pydantic = "^2.11.1"
pydantic-settings = "^2.10.1"

# Security / JWT
python-jose = {extras=["cryptography"], version = "^3.3.0"}
passlib = {extras=["bcrypt"], version = "^1.7.4"}

# Cache & messaging
redis = "^5.0.4"
aiokafka = "^0.12.0"   # switched to latest stable with PEP 517 support

# gRPC
grpcio = "^1.67.1"
grpcio-tools = "^1.74.0"

# Observability
prometheus-fastapi-instrumentator = "^6.0.0"
structlog = "^24.1.0"

# HTTP clients
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.8"
isort = "^5.13.2"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"